+++ src/draw/draw.h
@@ -6,7 +6,7 @@
 /*   By: tfujiwar <tfujiwar@student.42tokyo.jp>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/11/07 09:33:08 by tfujiwar          #+#    #+#             */
-/*   Updated: 2023/01/15 18:16:47 by tfujiwar         ###   ########.fr       */
+/*   Updated: 2023/01/15 18:48:07 by tfujiwar         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -17,7 +17,7 @@
 
 void	draw(t_env *env);
 bool	get_nearest_shape(const t_scene *scene, const t_ray *ray, \
-							float max_dist, bool exit_once_found, \
+							double max_dist, bool exit_once_found, \
 							t_shape **out_shape, t_intersect *out_intp);
 bool	raytrace(const t_scene *scene, const t_ray *eye_ray, t_rgb *rgb);
 void	pixel_put(t_env *env, int x, int y, int color);
